+++ src/app/(components)/create-post/error-modal.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { CustomError } from "@/api/lib/errors"
+import { UnknownError } from "@/components/ui/modal/components"
 import { ErrorModal } from "@/components/ui/modal/error-modal"
 
 // 基本的なエラーはこの関数から開ける
@@ -32,41 +33,9 @@ export function openErrorWithCode(error: CustomError, openModal: (c: React.React
   openModal(<UnknownError>{error.message}</UnknownError>)
 }
 
-// エラーの「詳細」の表示
-function Details({ title, children }: { 
-  title?: React.ReactNode,
-  children: React.ReactNode 
-}) {
-  return (
-    <div>
-      <h3 className="font-bold mb-3">{title ?? "詳細"}</h3>
-      <div className="border-red-600 border px-3 py-2">{children}</div>
-    </div>
-  )
-}
-
 export function ValidationError() {
   return (
     <p>入力された値が無効です</p>
-  )
-}
-
-export function UnknownError({ children }: { children: React.ReactNode }) {
-  return (
-    <ErrorModal
-      title="UNKNOWN ERROR!!"
-      label="不明なエラー"
-      message="予期せぬエラーが発生しました。"
-    >
-      <Details>{children}</Details>
-      <hr className="border-red-600 mt-6 mb-4" />
-      <p>
-        未解決のエラーが発生しました。<br />
-        このエラーを修正するため、
-        <a className="text-secondary-400" href="https://github.com/nanasi-1/encrypted-board">GitHub</a>
-        での貢献をご検討ください。<br />
-      </p>
-    </ErrorModal>
   )
 }
